Add volume instant command
