+++ .gitignore
@@ -50,4 +50,4 @@ pkg/loki/wal
 .vscode
 
 # nix
-result
+nix/result
