+++ pkg/logcli/query/query.go
@@ -7,15 +7,10 @@ import (
 	"io"
 	"log"
 	"os"
-	"sort"
-	"strings"
 	"sync"
-	"text/tabwriter"
 	"time"
 
-	"github.com/fatih/color"
 	"github.com/grafana/dskit/multierror"
-	json "github.com/json-iterator/go"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/weaveworks/common/user"
@@ -23,11 +18,10 @@ import (
 
 	"github.com/grafana/loki/pkg/logcli/client"
 	"github.com/grafana/loki/pkg/logcli/output"
+	"github.com/grafana/loki/pkg/logcli/print"
 	"github.com/grafana/loki/pkg/loghttp"
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/logql"
-	"github.com/grafana/loki/pkg/logqlmodel"
-	"github.com/grafana/loki/pkg/logqlmodel/stats"
 	"github.com/grafana/loki/pkg/loki"
 	"github.com/grafana/loki/pkg/storage"
 	chunk "github.com/grafana/loki/pkg/storage/chunk/client"
@@ -40,11 +34,6 @@ import (
 )
 
 const schemaConfigFilename = "schemaconfig"
-
-type streamEntryPair struct {
-	entry  loghttp.Entry
-	labels loghttp.LabelSet
-}
 
 // Query contains all necessary fields to execute instant and range queries and print the results.
 type Query struct {
@@ -127,15 +116,17 @@ func (q *Query) DoQuery(c client.Client, out output.LogOutput, statistics bool) 
 		out = out.WithWriter(partFile)
 	}
 
+	result := print.NewQueryResultPrinter(q.ShowLabelsKey, q.IgnoreLabelsKey, q.Quiet, q.FixedLabelsLen, q.Forward)
+
 	if q.isInstant() {
 		resp, err = c.Query(q.QueryString, q.Limit, q.Start, d, q.Quiet)
 		if err != nil {
 			log.Fatalf("Query failed: %+v", err)
 		}
 		if statistics {
-			q.printStats(resp.Data.Statistics)
+			result.PrintStats(resp.Data.Statistics)
 		}
-		_, _ = q.printResult(resp.Data.Result, out, nil)
+		_, _ = result.PrintResult(resp.Data.Result, out, nil)
 	} else {
 		unlimited := q.Limit == 0
 
@@ -164,10 +155,10 @@ func (q *Query) DoQuery(c client.Client, out output.LogOutput, statistics bool) 
 			}
 
 			if statistics {
-				q.printStats(resp.Data.Statistics)
+				result.PrintStats(resp.Data.Statistics)
 			}
 
-			resultLength, lastEntry = q.printResult(resp.Data.Result, out, lastEntry)
+			resultLength, lastEntry = result.PrintResult(resp.Data.Result, out, lastEntry)
 			// Was not a log stream query, or no results, no more batching
 			if resultLength <= 0 {
 				break
@@ -402,24 +393,6 @@ func maxTime(t1, t2 time.Time) time.Time {
 	return t2
 }
 
-func (q *Query) printResult(value loghttp.ResultValue, out output.LogOutput, lastEntry []*loghttp.Entry) (int, []*loghttp.Entry) {
-	length := -1
-	var entry []*loghttp.Entry
-	switch value.Type() {
-	case logqlmodel.ValueTypeStreams:
-		length, entry = q.printStream(value.(loghttp.Streams), out, lastEntry)
-	case loghttp.ResultTypeScalar:
-		q.printScalar(value.(loghttp.Scalar))
-	case loghttp.ResultTypeMatrix:
-		q.printMatrix(value.(loghttp.Matrix))
-	case loghttp.ResultTypeVector:
-		q.printVector(value.(loghttp.Vector))
-	default:
-		log.Fatalf("Unable to print unsupported type: %v", value.Type())
-	}
-	return length, entry
-}
-
 // DoLocalQuery executes the query against the local store using a Loki configuration file.
 func (q *Query) DoLocalQuery(out output.LogOutput, statistics bool, orgID string, useRemoteSchema bool) error {
 	var conf loki.Config
@@ -502,8 +475,9 @@ func (q *Query) DoLocalQuery(out output.LogOutput, statistics bool, orgID string
 		return err
 	}
 
+	resPrinter := print.NewQueryResultPrinter(q.ShowLabelsKey, q.IgnoreLabelsKey, q.Quiet, q.FixedLabelsLen, q.Forward)
 	if statistics {
-		q.printStats(result.Statistics)
+		resPrinter.PrintStats(result.Statistics)
 	}
 
 	value, err := marshal.NewResultValue(result.Data)
@@ -511,7 +485,7 @@ func (q *Query) DoLocalQuery(out output.LogOutput, statistics bool, orgID string
 		return err
 	}
 
-	q.printResult(value, out, nil)
+	resPrinter.PrintResult(value, out, nil)
 	return nil
 }
 
@@ -572,159 +546,6 @@ func (q *Query) SetInstant(time time.Time) {
 
 func (q *Query) isInstant() bool {
 	return q.Start == q.End && q.Step == 0
-}
-
-func (q *Query) printStream(streams loghttp.Streams, out output.LogOutput, lastEntry []*loghttp.Entry) (int, []*loghttp.Entry) {
-	common := commonLabels(streams)
-
-	// Remove the labels we want to show from common
-	if len(q.ShowLabelsKey) > 0 {
-		common = matchLabels(false, common, q.ShowLabelsKey)
-	}
-
-	if len(common) > 0 && !q.Quiet {
-		log.Println("Common labels:", color.RedString(common.String()))
-	}
-
-	if len(q.IgnoreLabelsKey) > 0 && !q.Quiet {
-		log.Println("Ignoring labels key:", color.RedString(strings.Join(q.IgnoreLabelsKey, ",")))
-	}
-
-	if len(q.ShowLabelsKey) > 0 && !q.Quiet {
-		log.Println("Print only labels key:", color.RedString(strings.Join(q.ShowLabelsKey, ",")))
-	}
-
-	// Remove ignored and common labels from the cached labels and
-	// calculate the max labels length
-	maxLabelsLen := q.FixedLabelsLen
-	for i, s := range streams {
-		// Remove common labels
-		ls := subtract(s.Labels, common)
-
-		if len(q.ShowLabelsKey) > 0 {
-			ls = matchLabels(true, ls, q.ShowLabelsKey)
-		}
-
-		// Remove ignored labels
-		if len(q.IgnoreLabelsKey) > 0 {
-			ls = matchLabels(false, ls, q.IgnoreLabelsKey)
-		}
-
-		// Overwrite existing Labels
-		streams[i].Labels = ls
-
-		// Update max labels length
-		len := len(ls.String())
-		if maxLabelsLen < len {
-			maxLabelsLen = len
-		}
-	}
-
-	// sort and display entries
-	allEntries := make([]streamEntryPair, 0)
-
-	for _, s := range streams {
-		for _, e := range s.Entries {
-			allEntries = append(allEntries, streamEntryPair{
-				entry:  e,
-				labels: s.Labels,
-			})
-		}
-	}
-
-	if len(allEntries) == 0 {
-		return 0, nil
-	}
-
-	if q.Forward {
-		sort.Slice(allEntries, func(i, j int) bool { return allEntries[i].entry.Timestamp.Before(allEntries[j].entry.Timestamp) })
-	} else {
-		sort.Slice(allEntries, func(i, j int) bool { return allEntries[i].entry.Timestamp.After(allEntries[j].entry.Timestamp) })
-	}
-
-	printed := 0
-	for _, e := range allEntries {
-		// Skip the last entry if it overlaps, this happens because batching includes the last entry from the last batch
-		if len(lastEntry) > 0 && e.entry.Timestamp == lastEntry[0].Timestamp {
-			skip := false
-			// Because many logs can share a timestamp in the unlucky event a batch ends with a timestamp
-			// shared by multiple entries we have to check all that were stored to see if we've already
-			// printed them.
-			for _, le := range lastEntry {
-				if e.entry.Line == le.Line {
-					skip = true
-				}
-			}
-			if skip {
-				continue
-			}
-		}
-		out.FormatAndPrintln(e.entry.Timestamp, e.labels, maxLabelsLen, e.entry.Line)
-		printed++
-	}
-
-	// Loki allows multiple entries at the same timestamp, this is a bit of a mess if a batch ends
-	// with an entry that shared multiple timestamps, so we need to keep a list of all these entries
-	// because the next query is going to contain them too and we want to not duplicate anything already
-	// printed.
-	lel := []*loghttp.Entry{}
-	// Start with the timestamp of the last entry
-	le := allEntries[len(allEntries)-1].entry
-	for i, e := range allEntries {
-		// Save any entry which has this timestamp (most of the time this will only be the single last entry)
-		if e.entry.Timestamp.Equal(le.Timestamp) {
-			lel = append(lel, &allEntries[i].entry)
-		}
-	}
-
-	return printed, lel
-}
-
-func (q *Query) printMatrix(matrix loghttp.Matrix) {
-	// yes we are effectively unmarshalling and then immediately marshalling this object back to json.  we are doing this b/c
-	// it gives us more flexibility with regard to output types in the future.  initially we are supporting just formatted json but eventually
-	// we might add output options such as render to an image file on disk
-	bytes, err := json.MarshalIndent(matrix, "", "  ")
-	if err != nil {
-		log.Fatalf("Error marshalling matrix: %v", err)
-	}
-
-	fmt.Print(string(bytes))
-}
-
-func (q *Query) printVector(vector loghttp.Vector) {
-	bytes, err := json.MarshalIndent(vector, "", "  ")
-	if err != nil {
-		log.Fatalf("Error marshalling vector: %v", err)
-	}
-
-	fmt.Print(string(bytes))
-}
-
-func (q *Query) printScalar(scalar loghttp.Scalar) {
-	bytes, err := json.MarshalIndent(scalar, "", "  ")
-	if err != nil {
-		log.Fatalf("Error marshalling scalar: %v", err)
-	}
-
-	fmt.Print(string(bytes))
-}
-
-type kvLogger struct {
-	*tabwriter.Writer
-}
-
-func (k kvLogger) Log(keyvals ...interface{}) error {
-	for i := 0; i < len(keyvals); i += 2 {
-		fmt.Fprintln(k.Writer, color.BlueString("%s", keyvals[i]), "\t", fmt.Sprintf("%v", keyvals[i+1]))
-	}
-	k.Flush()
-	return nil
-}
-
-func (q *Query) printStats(stats stats.Result) {
-	writer := tabwriter.NewWriter(os.Stderr, 0, 8, 0, '\t', 0)
-	stats.Log(kvLogger{Writer: writer})
 }
 
 func (q *Query) resultsDirection() logproto.Direction {
